Fixed up a warning about buffer overflow

+++ DX12/00 - DX12 Starter/DX12Helper.cpp
@@ -65,7 +65,7 @@ Microsoft::WRL::ComPtr<ID3D12Resource> DX12Helper::CreateStaticBuffer(unsigned i
 	desc.MipLevels = 1;
 	desc.SampleDesc.Count = 1;
 	desc.SampleDesc.Quality = 0;
-	desc.Width = dataStride * dataCount; // Size of the buffer
+	desc.Width = (UINT64)dataStride * (UINT64)dataCount; // Size of the buffer
 
 	device->CreateCommittedResource(
 		&props,
